==> sesion-07/2-manipulacion_de_strings.ipynb <==
## Ejemplo 2: Manipulación de Strings

### 1. Objetivos:
    - Aprender a usar `replace`, `strip`, `title`, `upper`, `lower` y `split` para transformar datos tipo `string``
 
### 2. Desarrollo:

import pandas as pd

Iniciamos leyendo el conjunto de datos, ahora usaremos el archivo `new_york_times_bestsellers-dirty.csv` o el obtenido del ejemplo anterior:

df = pd.read_csv("new_york_times_bestsellers-dirty.csv", index_col=0)
df.head(3)

Empecemos con la columna `description` que tiene un 'Descr:' al inicio de cada texto. Si queremos remover ese texto podemos usar el método `replace` de la propiedad `str` de esa `Serie` en la forma:

`dataframe[-columna-].str.replace("origen", "destino")`

df["description"] = df["description"].str.replace("Descr:", "")

df.head(3)

Podemos consultar una fila de la columna `description` para comprobar el resultado con:

`dataframe.loc[-fila-, -columna-]`

df.loc[0,"description"]

Ahora se observa también espacios vacíos al principio y final de nuestras textos o cadenas, vamos a removerlos usando `strip` que elimina espacios en blanco al inicio y al final en la forma:

`dataframe[-columna-].str.strip()`

df['description'] = df['description'].str.strip()

df.loc[0, "description"]

Perfecto, examinemos el DataFrame de nuevo:

df.head(3)

Ahora veamos la columna 'title', cuyos textos están en mayúsculas. Esto no es muy agradable, así que podemos usar algunos métodos para modificar el patrón de mayúsculas y minúsculas con la función `df.str.title()`:

df["title"] = df["title"].str.title()

df.head(3)

Ahora, digamos que queremos separar nuestra columna `author` en dos columnas `author_first_name` y `author_last_name`. Eso lo podemos hacer con el método `split` en la forma:

`dataframe[-columna-].str.split(-separador-)`

df["author"].str.split()

Podemos convertirlo en dos columnas así:

`dataframe[-columna-].str.split(-separador-, expand=True)`

df["author"].str.split(expand=True)

Y podemos asignar las nuevas columnas de forma compacta usando la forma:

`df[-lista de nuevas columnas-] = -dataframe con los valores de las nuevas columnas-`

df[["author_first_name","author_last_name"]] = df["author"].str.split(expand=True)

df.head(3)

¡Éxito!

---
---

## Reto 2: Manipulación de Strings

### 1. Objetivos:
    - Practicar manipular `strings` usando métodos como `split`, `title`, `strip, etc.
 
---
    
### 2. Desarrollo:

#### a) Limpiando texto

Vamos a trabajar en la versión del dataset que guardaste en el reto pasado. Las acciones que tienes que tomar en este Reto son las siguientes:

1. Reemplaza los guiones en las `strings` de la columna `orbit_class_description` por espacios.
2. Elimina los espacios vacíos al principio y final de las `strings` de la misma columna.
3. Hay una columna llamada `id_name` que contiene el 'id' y el nombre de cada objeto separados por un guión. Separa estos datos en dos columnas llamadas `id` y `name`.
4. Haz que las `strings` de la columna `orbiting_body` empiecen con mayúscula.
5. Asigna el `DataFrame` resultante a la variable `df_reto_2`.
6. Guarda tu resultado en un archivo .csv.

# import
import pandas as pd

df_reto_2 = pd.read_csv('near_earth_objects-jan_feb_1995-reto_1.csv')

df_reto_2.head(3)

df_reto_2.dtypes

# remplaza
df_reto_2["orbit_class_description"] = df_reto_2["orbit_class_description"].str.replace("-", " ")
df_reto_2

# reasigna
...

df_reto_2["orbit_class_description"].loc[0]

# elimina espacios en blanco
df_reto_2["orbit_class_description"] = df_reto_2['orbit_class_description'].str.strip()

df_reto_2["orbit_class_description"].loc[0]

# separa por guiones
df_reto_2[["id","name"]] = df_reto_2["id_name"].str.split("-",expand=True)

df_reto_2.head(3)

# cambia a títulos
df_reto_2["orbiting_body"] = df_reto_2["orbiting_body"].str.title()

df_reto_2["orbiting_body"]

# guarda en el archivo objetos_cercanos_2.csv
df_reto_2.to_csv('objetos_cercanos_2.csv')

¡Compara con el Ingeniero de Datos de tu confianza y si coinciden entonces envíen su respuesta!
