==> sesion-07/3-map.ipynb <==
## Ejemplo 3: `map`

### 1. Objetivos:
    - Usar `map` para convertir datos usando un `diccionario` o una `función``
 
---
    
### 2. Desarrollo:

import pandas as pd

Nuevemente puedes usar el archivo `new_york_times_bestsellers-dirty.csv` para usar los datos sin procesar:

df = pd.read_csv("new_york_times_bestsellers-dirty.csv" , index_col=0)

df.head(3)

Para imprimir todos los elementos únicos en la columna `rank.numberInt` se usa la siguiente función:

`dataframe[-columna-].unique()`

df["rank.numberInt"].unique()

Digamos que queremos transformar los datos de nuestra columna 'rank.numberInt' para que esté dado por letras, no por números.

Sabemos que hay un valor 'No Rank' en esa columna, así que nuestro diccionario de conversión podría verse así:

int_a_letra = {
    '1': 'a',
    '2': 'b',
    '3': 'c',
    '4': 'd',
    '5': 'e',
    '6': 'f',
    '7': 'g',
    '8': 'h',
    '9': 'i',
    '10': 'j',
    '11': 'k',
    '12': 'l',
    '13': 'm',
    '14': 'n',
    '15': 'o',
    '16': 'p',
    'No Rank': 'z'
}

Lo aplicamos usando `map` y podemos aplicar `head()` al mismo tiempo para cortar resultados:

`dataframe[-columna-].map(-diccionario-)`

df["rank.numberInt"].map(int_a_letra)

Y como sabemos, también podemos usar una función para `map`, por ejemplo para realizar una correspondencia entre el precio de un libro y su representación en `string`:

def numero_a_pesos_mn(valor):
    """ convierte de float a pesos en el formato de moneda nacional """
    return f"$ {valor} M.N."

df['price.numberDouble'].map(numero_a_pesos_mn).head(20)

Lo único que tienes que pensar al usar `map` es: "¿Este dato tiene una correspondencia con otro dato que pueda representar con un diccionario o una función?". Y listo.

---
---

## Reto 3: Map

### 1. Objetivos:
    - Practicar el uso del método `map` para mapear un dato a otro dato que le corresponde.
    
---
    
### 2. Desarrollo:

#### a) Booleanos a numéricos

Vamos a trabajar sobre el dataset que guardaste en el Reto anterior. Esta vez tu Reto es muy sencillo:

1. La columna `is_potentially_hazardous_asteroid` tiene valores `booleanos`. Crea un diccionario de mapeo donde hagas un correspondencia de cada valor `booleano` con su equivalente numérico y transforma esa columna.
2. Usa una función para mapear la columna `relative_velocity.kilometers_per_hour` a una nueva columna llamada `relative_velocity.kilometers_per_minute`, que contenga la velocidad del objeto en kilómetros por minuto.
3. Guarda tu `DataFrame` resultante en la variable `df_reto_3`.
4. Guarda tu resultado en un archivo .csv.

import pandas as pd

df_reto_3 = pd.read_csv("objetos_cercanos_2.csv", index_col=0)
df_reto_3.head(3)

# analiza
valores_booleanos = {
    True: 1,
    False: 0
}

#mapea
df_reto_3["is_potentially_hazardous_asteroid"] = df_reto_3["is_potentially_hazardous_asteroid"].map(valores_booleanos)

# mapea
df_reto_3
df_reto_3.dtypes

# analiza

# transforma usando función
def agregar_kmh(valor):
    return valor/60

df_reto_3["relative_velocity.kilometers_per_minute"] = df_reto_3['relative_velocity.kilometers_per_hour'].map(agregar_kmh)

df_reto_3

# guarda resultado al archivo objetos_cercanos_3.csv
df_reto_3.to_csv('objetos_cercanos_3.csv')

A continuación la celda de validación que en este caso sólo te ayuda a corroborar si estás realizando bien parte del proceso, compara tus resultados con tu Ingenier@ de Datos de confianza y si coinciden envía tus resultados

def revisar_resultados(df):
    
    import pandas as np
    import pandas.api.types as pdtypes
    
    assert pdtypes.is_int64_dtype(df['is_potentially_hazardous_asteroid']), 'La columna "is_potentially_hazardous_asteroid" no ha sido transformada a tipo numerico'
    assert len(df['is_potentially_hazardous_asteroid'].unique()) == 2, 'Hubo un error con la correspondencia de valores booleanos a numéricos. Hay más de dos valores posibles en la columna resultante'
    assert df['relative_velocity.kilometers_per_minute'].equals(df['relative_velocity.kilometers_per_hour'] / 60), 'La conversión de kilometros por hora a kilómetros por minuto no fue realizada correctamente'
    
    print(f'Todos los procesos fueron realizados exitosamente!')

revisar_resultados(df_reto_3)
