==> sesion-07/4-apply.ipynb <==
## Ejemplo 4: `apply`

### 1. Objetivos:
    - Usar `apply` para aplicar funciones a `Series` y `DataFrames`
 
---
    
### 2. Desarrollo:

import pandas as pd
import numpy as np

Para éste ejemplo también usamos el archivo `new_york_times_bestsellers-dirty.csv` porque vamos a aplicar una tranformación diferente a los mismos datos:

df = pd.read_csv('new_york_times_bestsellers-dirty.csv', index_col=0)
df.head(3)

Podemos aplicar funciones con uno o más parámetros a nuestras `Series` con el método `apply`, así que vamos a convertir la fecha de la columna `published_date.numberLong` que está dado en milisegundos en formato Epoc o Timestamp a sólo tener el año de publicación, recuerda usar:

`pd.to_datetime(-valor-, unit="ms")`

def anio_de_publicacion(valor):
    """ obtener el año desde la fecha epoc """
    fecha = pd.to_datetime(valor, unit="ms")
    return fecha.year

df['published_date.numberLong'].apply(anio_de_publicacion)

O esta otra, obtener la semanas en listas, pero en porcentaje en base al máximo, en este caso necesitamos pasar un valor adiciona a la función y se realiza de la forma:

`df[-columna-].apply(-función-, args=(-máximo de la columna-, ...) )`

primero creamos nuestra función:

def semanas_en_porcentaje(valor, maximo):
    """ Regresar el porcentaje en relación a maximo """
    pass

def semanas_en_porcentaje(valor, maximo):
    """ Regresar el porcentaje en relación a maximo """
    porcentaje = valor / maximo * 100
    return f"{porcentaje:.2f} %"

Y lo aplicamos a la columna `weeks_on_list.numberInt`, pero también necesitamos el valor máximo de esta columna, que lo podemos obtener con `serie.max()`, entonces la función `apply()` queda como sigue:

df['weeks_on_list.numberInt'].apply(semanas_en_porcentaje, args=(df["weeks_on_list.numberInt"].max(), ))

---
---

## Reto 4: Apply

### 1. Objetivos:
    - Practicar el uso del método `apply` para obtener columnas nuevas a partir de columnas existentes
    
### 2. Desarrollo:

#### a) Obteniendo columnas nuevas a partir de existentes

Vamos a trabajar con el dataset que guardaste de tu Reto anterior. Esta vez tu Reto es el siguiente:

1. Crea una función que reciba un valor (en este caso el diámetro en metros de un objeto espacial) y regrese la proporción de ese valor en comparación con el diámetro de la Tierra. El diámetro de la Tierra es de 12,742 km. Así que el diámetro de un objeto que mida 10000 metros corresponde a un valor de 0.00078 en proporción al diámetro de la Tierra.
2. Usa la columna 'estimated_diameter.meters.estimated_diameter_max', aplícale la función usando `apply` y crea una nueva columna llamada `proportion_of_max_diameter_to_earth` que deberá de ser de tipo `float64` y deberá contener valores entre 0 y 1 ya que es una proporción.
3. Asigna el resultado a la variable `df_reto_4`.
4. Guarda tu conjunto de datos en un archivo .csv.

df_reto_4 = pd.read_csv("objetos_cercanos_3.csv", index_col=0)
df_reto_4.head(3)

valor = (10000/1000)  / (12742)  
valor
df_reto_4.dtypes
df_reto_4

def proporcion_vs_tierra(valor):
    """ Regresa la proporción de valor con respecto a la Tierra """
    proporcion = (valor/ 12742000)
    
    return proporcion

df_reto_4["proportion_of_max_diameter_to_earth"] = df_reto_4["estimated_diameter.meters.estimated_diameter_max"].apply(proporcion_vs_tierra)
df_reto_4["proportion_of_max_diameter_to_earth"]

# guarda en el archivo objetos_cercanos_4.csv
df_reto_4.to_csv('objetos_cercanos_4.csv')

La celda de validación nuevamente te ayudará a darte pautas para saber si tus resultados van por buen camino!

def revisar_aplicacion(df):
    
    assert 'proportion_of_max_diameter_to_earth' in df, 'No existe una columna llamada "proportion_of_max_diameter_to_earth" en el DataFrame'
    assert df['proportion_of_max_diameter_to_earth'].equals(df['estimated_diameter.meters.estimated_diameter_max'] / 12742000), 'La transformacion no fue realizada adecuadamente'
    
    print(f'La transformación y creación de una nueva columna fue realizada exitosamente!')

revisar_aplicacion(df_reto_4)
