==> sesion-07/6-sort.ipynb <==
## Ejemplo 6: Sort

### 1. Objetivos:
    - Aprender a usar `sort_values` para reordenar nuestros datos
 
### 2. Desarrollo:

import pandas as pd

df = pd.read_csv('new_york_times_bestsellers-dirty.csv', index_col=0)
df.head(3)

Finalmente veamos como podemos ordenar los elementos de una DataFrame, por ejemplo, queremos ordenar nuestras entradas empezando por el libro de mayor precio hasta el libro de menor precio, así que se ordena en base a la columna `price.numberDouble` de manera descendente usando la siguiente forma de la función `sort_values()`:

`dataframe.sort_values(-columna-, ascending=False)`

df.sort_values("price.numberDouble", ascending=False)

Si ordenamos en base a `published_date.numberLong` podemos también obtener la lista desde la publicación más antigua hasta la publicación más reciente:

df.sort_values("published_date.numberLong")

Pero como seguramente no lo crees aún porque estamos usando la columna del Timestamp, vamos a convertir la fechas a datetime con el método:

`pd.to_datetime(-dataframe[columna]-, unit="ms")`

y aplicarlo desde el DataFrame original, ordenamos nuevamente y obtenemos un DataFrame sólo con las columnas `amazon_product_url` y `published_date.numberLong` para simplificar el resultado:

df["published_date.numberLong"] = pd.to_datetime(df["published_date.numberLong"], unit="ms")

df.head()

df_ordenado = df.sort_values("published_date.numberLong")

df_ordenado[[ "amazon_product_url","published_date.numberLong"]]

**NOTA PYTHONESCA:** Para trabajar con fechas se pueden usar los tipos de datos `datetime` o `int` (en alguna unidad de segundos), el ordenamiento u otra operación se realizará de forma correcta, pero procesar fechas de tipo `object` o `str` podríamos obtener resultados inesperados.

Por ejemplo, podríamos primero filtrar para sólo tener los libros de la editorial que tiene más libros como 'best sellers' usando la columna `publisher` y después ordenarlos del que pasó más días en la lista de 'best sellers' al que pasó menos días en la lista usando la columna `weeks_on_list.numberInt`.

Primero encontramos la editorial con más bestsellers con el método:

`dataframe[-columna-].value_counts()`

df["publisher"].value_counts()

Filtra sólo los que perteneces a la editorial encontrada (Putnam?) usando la aplicación de filtros con:

`dataframe[ dataframe[-columna-] op -valor- ]`

df_1 =  df[ df["publisher"]=="Putnam" ]
df_1.head(3)

Ordenando en base a la cantidad de días como bestsellers usando:

`dataframe.sort_values(-columna-, ascending=False)`

df.sort_values("weeks_on_list.numberInt", ascending=False)

---
---

## Reto 6: Sort

### 1. Objetivos:
    - Practicar el uso de `sort_values` para obtener datos específicos
    
### 2. Desarrollo:

#### a) Reordenamiento para hallazgo de valores

Vamos a trabajar sobre el dataset que guardaste en la sesión anterior. Tu Reto consiste en lo siguiente:

1. Usa `sort_values` y `loc` para obtener el valor de la velocidad en kilómetros por segundo más lenta de todos los objetos que tenemos en el dataset y asígnalo a `velocidad_en_kilometros_por_segundo_de_objeto_mas_lento`.
2. Usa `sort_values` y `loc` para obtener el valor del diámetro más grande medido que existe en nuestro dataset y asígnalo a `medida_de_diametro_mas_grande`

df_reto_6 = pd.read_csv("objetos_cercanos_4.csv", index_col=0)
df_reto_6.head(3)

df_ordenado=df_reto_6.sort_values(by="relative_velocity.kilometers_per_second", ascending=True)
#velocidad_en_kilometros_por_segundo_de_objeto_mas_lento = df_ordenado.loc[0]
velocidad_en_kilometros_por_segundo_de_objeto_mas_lento = df_ordenado.iloc[0]["relative_velocity.kilometers_per_second"]
velocidad_en_kilometros_por_segundo_de_objeto_mas_lento


df_ordenado=df_reto_6.sort_values("estimated_diameter.meters.estimated_diameter_max", ascending=False)
#velocidad_en_kilometros_por_segundo_de_objeto_mas_lento = df_ordenado.loc[0]
medida_de_diametro_mas_grande =  df_ordenado.iloc[0]["estimated_diameter.meters.estimated_diameter_max"]
medida_de_diametro_mas_grande



En este caso la celda de valición si comprueba si tus resultados son correctos, adelante ...

def corroborar_hallazgos(velocidad_en_kilometros_por_segundo_de_objeto_mas_lento, medida_de_diametro_mas_grande):
    
    assert velocidad_en_kilometros_por_segundo_de_objeto_mas_lento == 0.681436673, 'Esa no es la velocidad en kilómetros por segundo del objeto más lento'
    assert medida_de_diametro_mas_grande == 6516.883821679, 'Ese no es el diámetro más grande medido en nuestro dataset'
    
    print('Tus hallazgos son correctos. ¡Bien hecho!')
    
corroborar_hallazgos(velocidad_en_kilometros_por_segundo_de_objeto_mas_lento, medida_de_diametro_mas_grande)
